// WADNR.Web/src/app/shared/generated/model/project-classification-detail-item.ts
/**
 * WADNR.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class ProjectClassificationDetailItem { 
    ProjectClassificationID?: number;
    ClassificationID?: number;
    ClassificationName?: string | null;
    ClassificationSystemID?: number;
    ClassificationSystemName?: string | null;
    ProjectClassificationNotes?: string | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface ProjectClassificationDetailItemForm { 
    ProjectClassificationID?: FormControl<number>;
    ClassificationID?: FormControl<number>;
    ClassificationName?: FormControl<string>;
    ClassificationSystemID?: FormControl<number>;
    ClassificationSystemName?: FormControl<string>;
    ProjectClassificationNotes?: FormControl<string>;
}

export class ProjectClassificationDetailItemFormControls { 
    public static ProjectClassificationID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static ClassificationID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static ClassificationName = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static ClassificationSystemID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static ClassificationSystemName = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static ProjectClassificationNotes = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}