// WADNR.Web/src/app/shared/generated/model/project-classification-save-request.ts
/**
 * WADNR.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ProjectClassificationItemRequest } from './project-classification-item-request';


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class ProjectClassificationSaveRequest { 
    Classifications?: Array<ProjectClassificationItemRequest> | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface ProjectClassificationSaveRequestForm { 
    Classifications?: FormControl<Array<ProjectClassificationItemRequest>>;
}

export class ProjectClassificationSaveRequestFormControls { 
    public static Classifications = (value: FormControlState<Array<ProjectClassificationItemRequest>> | Array<ProjectClassificationItemRequest> = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<Array<ProjectClassificationItemRequest>>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}